fix(simulation): review fixes for detailed analysis schemas and router
- Change DetailedNoteData.helpfulness_score from int to float to match
  computed averages from scoring
- Create typed DetailedAnalysisMeta schema replacing dict[str, Any] meta,
  making RequestVarianceMeta visible in OpenAPI/TypeScript
- Skip compute_request_variance on page > 1 (only needed on first page)
- Use loaders.ratings() instead of loaders.full() in variance computation
  since only ratings are needed, not request+archive data
- Simplify redundant request_ids union in compute_request_variance
- Fix inst_id fallback: return empty string instead of silently
  substituting user_profile_id when agent instance mapping is missing

+++ opennotes-cli/src/opennotes_cli/api_client/models/detailed_note_data.py
@@ -25,7 +25,7 @@ class DetailedNoteData:
         summary (str):
         classification (str):
         status (str):
-        helpfulness_score (int):
+        helpfulness_score (float):
         author_agent_name (str):
         author_agent_instance_id (str):
         request_id (None | str | Unset):
@@ -37,7 +37,7 @@ class DetailedNoteData:
     summary: str
     classification: str
     status: str
-    helpfulness_score: int
+    helpfulness_score: float
     author_agent_name: str
     author_agent_instance_id: str
     request_id: None | str | Unset = UNSET
